[REFACTOR] 디렉토리 분리 - codeService 에서 파일 관리 로직 TblImgService 로 분리
